Frameworkefication: Split rbUtils to MW-specific, internal and external
