+++ lib/auth.js
@@ -1,8 +1,8 @@
 "use strict";
 
-var P       = require('bluebird');
-var rbUtil  = require('./rbUtil');
-var URI     = require('swagger-router').URI;
+var P          = require('bluebird');
+var URI        = require('swagger-router').URI;
+var HTTPError  = require('./exports').HTTPError;
 
 
 /**
@@ -20,6 +20,27 @@ function constructInternalRequestRegex(reqWhiteList) {
     }).join('|');
     return internalReqRegex && internalReqRegex.length > 0
         ? new RegExp(internalReqRegex) : undefined;
+}
+
+/**
+ * Copies forwarded headers from restbase to request.
+ * If the same header was already set it takes precedence over
+ * the forwarded header.
+ *
+ * @param restbase a restbase instance to take the forwarded headers from
+ * @param req a request where to copy the headers
+ * @param headers array of header names to copy
+ */
+function copyForwardedHeaders(restbase, req, headers) {
+    if (restbase._rootReq && restbase._forwardedHeaders) {
+        req.headers = req.headers || {};
+        headers.forEach(function(header) {
+            if (!req.headers[header] && restbase._forwardedHeaders[header]) {
+                req.headers[header] = restbase._forwardedHeaders[header];
+            }
+        });
+    }
+    return req;
 }
 
 /**
@@ -45,7 +66,7 @@ var securityDefinitionCreators = {
         return {
             prepareRequest: function(restbase, req) {
                 if (isInternalRequest(req)) {
-                    rbUtil.copyForwardedHeaders(restbase, req,
+                    copyForwardedHeaders(restbase, req,
                         ['cookie', 'x-forwarded-for', 'x-client-ip']);
                 }
             },
@@ -77,7 +98,7 @@ var securityDefinitionCreators = {
                     if (userInfo && userInfo.rights && Array.isArray(userInfo.rights)) {
                         permissions.forEach(function(perm) {
                             if (userInfo.rights.indexOf(perm) < 0) {
-                                throw new rbUtil.HTTPError({
+                                throw new HTTPError({
                                     status: 401,
                                     body: {
                                         type: 'unauthorized',
@@ -88,7 +109,7 @@ var securityDefinitionCreators = {
                             }
                         });
                     } else {
-                        throw new rbUtil.HTTPError({
+                        throw new HTTPError({
                             status: 400,
                             body: {
                                 type: 'invalid_request',
@@ -137,7 +158,7 @@ var securityDefinitionCreators = {
                         }
 
                         if (!whitelistMap[patternName].test(headerValue)) {
-                            throw new rbUtil.HTTPError({
+                            throw new HTTPError({
                                 status: 403,
                                 body: {
                                     type: 'forbidden',
