+++ lib/handlerTemplate.js
@@ -6,11 +6,7 @@
  */
 
 var P = require('bluebird');
-var swaggerRouter = require('swagger-router');
-var rbUtil = require('../lib/rbUtil');
-
-var URI = swaggerRouter.URI;
-var Template = swaggerRouter.Template;
+var Template = require('swagger-router').Template;
 
 /**
  * Creates a JS function that verifies property equality
