+++ lib/restbase.js
@@ -6,11 +6,32 @@
 
 var jwt = require('jsonwebtoken');
 var P = require('bluebird');
-var rbUtil = require('./rbUtil');
-var HTTPError = rbUtil.HTTPError;
+var utils = require('./utils');
+var HTTPError = require('./exports').HTTPError;
 var preq = require('preq');
 var swaggerUI = require('./swaggerUI');
 var AuthService = require('./auth');
+
+
+/**
+ * Create a uniform but shallow request object copy with sane defaults. This
+ * keeps code dealing with this request monomorphic (good for perf), and
+ * avoids subtle bugs when requests shared between recursive requests are
+ * mutated in another control branch. At the very minimum, we are mutating the
+ *
+ * @param req original request object
+ * @returns a shallow copy of a provided requests
+ */
+function cloneRequest(req) {
+    return {
+        uri: req.uri || req.url || null,
+        method: req.method || 'get',
+        headers: req.headers || {},
+        query: req.query || {},
+        body: req.body !== undefined ? req.body : null,
+        params: req.params || {}
+    };
+}
 
 function RESTBase(options, req) {
     if (options && options.constructor === RESTBase) {
@@ -20,7 +41,7 @@ function RESTBase(options, req) {
         this.metrics = par.metrics;
         this.reqId = par.reqId ||
             req && req.headers && req.headers['x-request-id'] ||
-            rbUtil.generateRequestId();
+            utils.generateRequestId();
 
         this._parent = par;
         // Remember the request that led to this child instance at each level, so
@@ -301,6 +322,9 @@ RESTBase.prototype._request = function(req) {
 
     // Make sure we have a sane & uniform request object that doesn't change
     // (at least at the top level) under our feet.
+    var childReq = cloneRequest(req);
+
+    var match = priv.router.route(childReq.uri);
     var childReq = rbUtil.cloneRequest(req);
     var match = this._priv.router.route(childReq.uri);
     var methods = match && match.value && match.value.methods;
@@ -433,8 +457,7 @@ methods.forEach(function(method) {
 // @param {string} token
 // @return {string} JWT signed token
 RESTBase.prototype.encodeToken = function(token) {
-    var newToken = jwt.sign({ next: token }, this._priv.options.conf.salt);
-    return newToken;
+    return jwt.sign({ next: token }, this._priv.options.conf.salt);
 };
 
 // Decode signed token and decode the orignal token
