+++ lib/router.js
@@ -3,7 +3,6 @@
 var P = require('bluebird');
 var yaml = require('js-yaml');
 var fs = P.promisifyAll(require('fs'));
-var rbUtil = require('./rbUtil');
 var handlerTemplate = require('./handlerTemplate');
 var swaggerRouter = require('swagger-router');
 var Validator = require('./validator');
