+++ lib/server.js
@@ -8,15 +8,97 @@
  */
 
 var P = require('bluebird');
-
-var rbUtil = require('./rbUtil');
-var Restbase = require('./restbase');
+var Busboy = require('busboy');
+var qs = require('querystring');
+var url = require('url');
 var http = require('http');
-var fs = require('fs');
-var Router = require('./router');
-var URI = require('swagger-router').URI;
 var zlib = require('zlib');
 var stream = require('stream');
+
+var URI = require('swagger-router').URI;
+
+var exports = require('./exports');
+var Restbase = require('./restbase');
+var Router = require('./router');
+var utils = require('./utils');
+
+
+// Should make it into 0.12, see https://github.com/joyent/node/pull/7878
+var SIMPLE_PATH = /^(\/(?!\/)[^\?#\s]*)(?:\?([^#\s]*))?$/;
+function parseURL(uri) {
+    // Fast path for simple path uris
+    var fastMatch = SIMPLE_PATH.exec(uri);
+    if (fastMatch) {
+        return {
+            protocol: null,
+            slashes: null,
+            auth: null,
+            host: null,
+            port: null,
+            hostname: null,
+            hash: null,
+            search: fastMatch[2] || '',
+            pathname: fastMatch[1],
+            path: fastMatch[1],
+            query: fastMatch[2] && qs.parse(fastMatch[2]) || {},
+            href: uri
+        };
+    } else {
+        return url.parse(uri, true);
+    }
+}
+
+
+/**
+ * Parse a POST request into request.body with BusBoy
+ * Drops file uploads on the floor without creating temporary files
+ *
+ * @param {Request} req HTTP request
+ * @returns {Promise<>}
+ */
+function read(req) {
+    return new P(function(resolve) {
+        var chunks = [];
+        req.on('data', function(chunk) {
+            chunks.push(chunk);
+        });
+
+        req.on('end', function() {
+            resolve(Buffer.concat(chunks));
+        });
+    });
+}
+
+function parsePOST(req) {
+    var readIt = (req.method === 'PUT') ||
+        (req.method === 'POST' && req.headers &&
+            (/^application\/json/i.test(req.headers['content-type'])
+            || !req.headers['content-type']));
+
+    if (readIt) {
+        return read(req);
+    } else if (req.method !== 'POST') {
+        return P.resolve();
+    } else {
+        // Parse the POST
+        return new P(function(resolve) {
+            // Parse POST data
+            var bboy = new Busboy({
+                headers: req.headers,
+                // Increase the form field size limit from the 1M default.
+                limits: { fieldSize: 15 * 1024 * 1024 }
+            });
+            var body = {};
+            bboy.on('field', function(field, val) {
+                body[field] = val;
+            });
+            bboy.on('finish', function() {
+                resolve(body);
+            });
+            req.pipe(bboy);
+        });
+    }
+}
 
 function handleResponse(opts, req, resp, response) {
     if (response && response.status) {
@@ -42,7 +124,7 @@ function handleResponse(opts, req, resp, response) {
         rh['x-content-type-options'] = 'nosniff';
         rh['x-frame-options'] = 'SAMEORIGIN';
 
-        rbUtil.addCSPHeaders(response, { domain: req.params && req.params.domain });
+        exports.misc.addCSPHeaders(response, { domain: req.params && req.params.domain });
 
         // Propagate the request id header
         rh['x-request-id'] = opts.reqId;
@@ -178,7 +260,7 @@ function handleResponse(opts, req, resp, response) {
 function handleRequest(opts, req, resp) {
     // Set the request ID early on for external requests
     req.headers = req.headers || {};
-    req.headers['x-request-id'] = req.headers['x-request-id'] || rbUtil.generateRequestId();
+    req.headers['x-request-id'] = req.headers['x-request-id'] || utils.generateRequestId();
 
     var remoteAddr = req.headers['x-client-ip'] || req.socket.remoteAddress;
     req.headers['x-client-ip'] = remoteAddr;
@@ -216,7 +298,7 @@ function handleRequest(opts, req, resp) {
     }
 
     // Create a new, clean request object
-    var urlData = rbUtil.parseURL(req.url);
+    var urlData = parseURL(req.url);
 
     var newReq = {
         uri: new URI(urlData.pathname),
@@ -227,7 +309,7 @@ function handleRequest(opts, req, resp) {
 
     // Start off by parsing any POST data with BusBoy
     return rbUtil.parsePOST(req)
-    .catchThrow(new rbUtil.HTTPError({
+    .catchThrow(new exports.HTTPError({
         status: 400,
         body: {
             type: 'invalid_request'
@@ -265,7 +347,7 @@ function handleRequest(opts, req, resp) {
         if (!e || e.name !== 'HTTPError') {
             var originalError = e;
             var stack = e && e.stack;
-            e = new rbUtil.HTTPError({
+            e = new exports.HTTPError({
                 status: 500,
                 body: {
                     type: 'internal_error',
