+++ lib/swaggerUI.js
@@ -2,7 +2,6 @@
 
 var P = require('bluebird');
 var fs = P.promisifyAll(require('fs'));
-var path = require('path');
 // Swagger-ui helpfully exports the absolute path of its dist directory
 var docRoot = require('swagger-ui').dist + '/';
 var rbUtil = require('./rbUtil');
