+++ lib/validator.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var HTTPError = require('./rbUtil').HTTPError;
+var HTTPError = require('./exports').HTTPError;
 var constructAjv = require('ajv');
 
 /**
